update Dockerfile - keeping stuff clean
Adopting the cleaning methode from Emiel to ensure we have a more clean and slim docker image to work with. This change will do a clean of yum and explicitly removing content from /var/tmp and /var/cache/yum to be sure we do not have useless junk in the container and keep things clean and tight.